refactor: fix typo and unify the packages names
