+++ design_system/src/commonMain/kotlin/net/thechance/mena/designsystem/presentation/component/button/radioButton/RadioButtonDefaults.kt
@@ -1,4 +1,4 @@
-package net.thechance.mena.designsystem.presentation.component.button.radio_button
+package net.thechance.mena.designsystem.presentation.component.button.radioButton
 
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.runtime.Composable
