+++ design_system/src/commonMain/kotlin/net/thechance/mena/designsystem/presentation/component/dialog/Dialog.kt
@@ -1,6 +1,6 @@
 @file:OptIn(ExperimentalComposeUiApi::class)
 
-package net.thechance.mena.designsystem.presentation.component.dialogue
+package net.thechance.mena.designsystem.presentation.component.dialog
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
